Tundra2 lc

+++ src/Application/MumbleVoipModule/MicrophoneAdjustmentWidget.cpp
@@ -59,7 +59,7 @@ namespace MumbleVoip
 
         int bytes_per_frame = SAMPLES_IN_FRAME*SAMPLE_WIDTH/8;
         PCMAudioFrame frame(SAMPLE_RATE, SAMPLE_WIDTH, NUMBER_OF_CHANNELS, bytes_per_frame );
-        while (framework_->Audio()->GetRecordedSoundSize() > bytes_per_frame)
+        while((int)framework_->Audio()->GetRecordedSoundSize() > bytes_per_frame)
         {
             int bytes = framework_->Audio()->GetRecordedSoundData(frame.DataPtr(), bytes_per_frame);
             if (bytes != bytes_per_frame)
